@@ -13,6 +13,7 @@
   "files": [
     "/dist"
   ],
+  "type": "module",
   "author": "Eren Yenigul",
   "license": "ISC",
   "dependencies": {
@@ -27,7 +28,7 @@
     "ts-loader": "^9.4.4",
     "tsc": "^2.0.4",
     "tslib": "^2.6.2",
-    "typescript": "^5.2.2"
+    "typescript": "^5.5.4"
   },
   "devDependencies": {
     "@types/jest": "^29.5.4",
